+++ code/include/TslSocket.hpp
@@ -54,7 +54,7 @@ public:
    * @param maxLength max number of bytes to read. Usually size of buffer.
    * @return number of bytes read.
    */
-  virtual int read(char *buffer, int maxLength);
+  virtual size_t read(char *buffer, size_t maxLength);
 
 private:
   /**
